chore: move async load config
