+++ src/cli/commands/migrate.ts
@@ -1,8 +1,7 @@
 import { createConnection } from 'typeorm';
 import { normalize, resolve } from 'path';
 import { Connection } from 'typeorm/connection/Connection';
-import { buildRegexpIfValid, ConfigModule, Logger } from '../../core';
-import { asyncLoadConfig } from '../utils/async-load-config';
+import { asyncLoadConfig, buildRegexpIfValid, ConfigModule, Logger } from '../../core';
 
 const logger = Logger.contextualize('Migrate command', 'MEDEX-CLI');
