--- src/cli/utils/fixtures/default-case-non-async-data.js
@@ -1,10 +1,0 @@
-module.exports = {
-    projectConfig: {
-      database_database: "./medusa-db.sql",
-      database_type: "sqlite",
-      database_password:"password",
-      store_cors: "STORE_CORS",
-      admin_cors: "ADMIN_CORS",
-    },
-    plugins:[],
-  };
