+++ src/core/tests/fixtures/async-load-config/async-function-with-non-async-data.js
@@ -1,0 +1,14 @@
+async function asyncConfig() {
+	return {
+		projectConfig: {
+			database_type: 'sqlite',
+			database_database: './medusa-db.sql',
+			admin_cors: 'ADMIN_CORS',
+			store_cors: 'STORE_CORS',
+			database_password: 'password',
+		},
+		plugins: [],
+	};
+}
+
+module.exports = asyncConfig();
